+++ cocos/2d/CCActionPageTurn3D.cpp
@@ -74,7 +74,7 @@ void PageTurn3D::update(float time)
     float deltaTheta = sqrtf(time);
     float theta = deltaTheta > 0.5f ? (float)M_PI_2*deltaTheta : (float)M_PI_2*(1-deltaTheta);
     
-    float rotateByYAxis = (2-time)* M_PI;
+    float rotateByYAxis = (float)((2.0f-time)* M_PI);
     
     float sinTheta = sinf(theta);
     float cosTheta = cosf(theta);
