+++ cocos/2d/CCFontAtlas.cpp
@@ -52,7 +52,7 @@ FontAtlas::FontAtlas(Font &theFont)
     _fontFreeType = dynamic_cast<FontFreeType*>(_font);
     if (_fontFreeType)
     {
-        _lineHeight = _font->getFontMaxHeight();
+        _lineHeight = (float)_font->getFontMaxHeight();
         _fontAscender = _fontFreeType->getFontAscender();
         _letterEdgeExtend = 2;
 
@@ -399,7 +399,7 @@ bool FontAtlas::prepareLetterDefinitions(const std::u32string& utf32Text)
                 _currentPageOrigX = 0;
                 if (_currentPageOrigY + _lineHeight + _letterPadding + _letterEdgeExtend >= CacheTextureHeight)
                 {
-                    updateTextureContent(pixelFormat, startY);
+                    updateTextureContent(pixelFormat, (int)startY);
 
                     startY = 0.0f;
 
@@ -428,7 +428,7 @@ bool FontAtlas::prepareLetterDefinitions(const std::u32string& utf32Text)
             {
                 _currLineHeight = glyphHeight;
             }
-            _fontFreeType->renderCharAt(_currentPageData, _currentPageOrigX + adjustForExtend, _currentPageOrigY + adjustForExtend, bitmap, bitmapWidth, bitmapHeight);
+            _fontFreeType->renderCharAt(_currentPageData, (int)_currentPageOrigX + adjustForExtend, (int)_currentPageOrigY + adjustForExtend, bitmap, bitmapWidth, bitmapHeight);
 
             tempDef.U = _currentPageOrigX;
             tempDef.V = _currentPageOrigY;
@@ -461,7 +461,7 @@ bool FontAtlas::prepareLetterDefinitions(const std::u32string& utf32Text)
         _letterDefinitions[it.first] = tempDef;
     }
 
-    updateTextureContent(pixelFormat, startY);
+    updateTextureContent(pixelFormat, (int)startY);
     return true;
 }
 
@@ -471,7 +471,7 @@ void FontAtlas::updateTextureContent(backend::PixelFormat format, int startY)
     auto outlineSize = _fontFreeType->getOutlineSize();
     if (outlineSize > 0 && format == backend::PixelFormat::AI88)
     {
-        int nLen = CacheTextureWidth * (_currentPageOrigY - startY + _currLineHeight);
+        int nLen = int(CacheTextureWidth * (_currentPageOrigY - startY + _currLineHeight));
         data = _currentPageData + CacheTextureWidth * (int)startY * 2;
         memset(_currentPageDataRGBA, 0, 4 * nLen);
         for (auto i = 0; i < nLen; i++)
@@ -479,12 +479,12 @@ void FontAtlas::updateTextureContent(backend::PixelFormat format, int startY)
             _currentPageDataRGBA[i*4] = data[i*2];
             _currentPageDataRGBA[i*4+3] = data[i*2+1];
         }
-        _atlasTextures[_currentPage]->updateWithData(_currentPageDataRGBA, 0, startY, CacheTextureWidth, _currentPageOrigY - startY + _currLineHeight);
+        _atlasTextures[_currentPage]->updateWithData(_currentPageDataRGBA, 0, startY, CacheTextureWidth, (int)_currentPageOrigY - startY + _currLineHeight);
     }
     else
     {
         data = _currentPageData + CacheTextureWidth * (int)startY;
-       _atlasTextures[_currentPage]->updateWithData(data, 0, startY, CacheTextureWidth, _currentPageOrigY - startY + _currLineHeight);
+       _atlasTextures[_currentPage]->updateWithData(data, 0, startY, CacheTextureWidth, (int)_currentPageOrigY - startY + _currLineHeight);
     }
 }
