+++ cocos/2d/CCFontAtlasCache.cpp
@@ -71,7 +71,7 @@ FontAtlas* FontAtlasCache::getFontAtlasTTF(const _ttfConfig* config)
     if ( it == _atlasMap.end() )
     {
         auto font = FontFreeType::create(realFontFilename, config->fontSize, config->glyphs,
-            config->customGlyphs, useDistanceField, config->outlineSize);
+            config->customGlyphs, useDistanceField, (float)config->outlineSize);
         if (font)
         {
             auto tempAtlas = font->createFontAtlas();
