+++ cocos/2d/CCTileMapAtlas.cpp
@@ -164,12 +164,12 @@ Color3B TileMapAtlas::getTileAt(const Vec2& position) const
 
 void TileMapAtlas::updateAtlasValueAt(const Vec2& pos, const Color3B& value, int index)
 {
-    CCASSERT( index >= 0 && index < _textureAtlas->getCapacity(), "updateAtlasValueAt: Invalid index");
+    CCASSERT( index >= 0 && (size_t)index < _textureAtlas->getCapacity(), "updateAtlasValueAt: Invalid index");
 
     V3F_C4B_T2F_Quad* quad = &((_textureAtlas->getQuads())[index]);
 
-    int x = pos.x;
-    int y = pos.y;
+    int x = (int)pos.x;
+    int y = (int)pos.y;
     float row = (float) (value.r % _itemsPerRow);
     float col = (float) (value.r / _itemsPerRow);
